unify test/support/time_helpers with test/support/mock_time. Should have been the same file from the beginning
